Removing all rhai, repalcing with lua

+++ src/runes/attack.rs
@@ -31,10 +31,10 @@ impl Rune for Attack {
 
         let dr_1 = DamageRune::new(self.source_uid,
                                    self.target_uid,
-                                   attacker.get_base_attack() as i64);
+                                   attacker.get_base_attack());
         let dr_2 = DamageRune::new(self.source_uid,
                                    self.source_uid,
-                                   defender.get_base_attack() as i64);
+                                   defender.get_base_attack());
 
         game_state.add_to_attacked_this_turn(self.source_uid);
