+++ src/runes/heal.rs
@@ -9,13 +9,13 @@ use hlua;
 pub struct Heal {
     target_uid: UID,
     source_uid: UID,
-    amount: i64,
+    amount: u32,
 }
 
 implement_for_lua!(Heal, |mut _metatable| {});
 
 impl Heal {
-    pub fn new(target_uid: UID, source_uid: UID, amount: i64) -> Heal {
+    pub fn new(target_uid: UID, source_uid: UID, amount: u32) -> Heal {
         Heal {
             target_uid: target_uid,
             source_uid: source_uid,
@@ -26,7 +26,7 @@ impl Heal {
 
 impl Rune for Heal {
     fn execute_rune(&self, game_state: &mut GameState) {
-        let m_h = ModifyHealth::new(self.target_uid, self.amount);
+        let m_h = ModifyHealth::new(self.target_uid, self.amount as i32);
         game_state.execute_rune(Box::new(m_h));
     }
