+++ src/runes/set_health.rs
@@ -7,13 +7,13 @@ use hlua;
 #[derive(RustcDecodable, RustcEncodable, Clone, Debug)]
 pub struct SetHealth {
     card_uid: UID,
-    amount: i64,
+    amount: u32,
 }
 
 implement_for_lua!(SetHealth, |mut _metatable| {});
 
 impl SetHealth {
-    pub fn new(card_uid: UID, amount: i64) -> SetHealth {
+    pub fn new(card_uid: UID, amount: u32) -> SetHealth {
         SetHealth {
             card_uid: card_uid,
             amount: amount,
@@ -23,7 +23,13 @@ impl SetHealth {
 
 impl Rune for SetHealth {
     fn execute_rune(&self, mut game_state: &mut GameState) {
+        //TODO: Make this work the way it is supposed to 
+
+        
         game_state.get_mut_minion(self.card_uid).unwrap().set_total_health(self.amount);
+    
+    
+    
     }
 
     fn can_see(&self, _controller: UID, _game_state: &GameState) -> bool {
